Create an Account and Login
Create an Account by first entering your username, then the password. Password should have at least 8 literals long, if not, it will ask again to enter a password.
Once the Account is created, it asks you to log in to your account. You have to log on to your account using the previously created Username and Password.
If the entered creditionals are wrong, it prompts and makes you enter the value again. 
Once the correct creditionals are entered, you are logged on to your Account. ; )